# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Game backend API với PlayFab integration, anti-cheat validation, và talent system",
        "title": "Hollow Wilds Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gamefeel.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.1.0"
    },
    "host": "gamefeel-backend.fly.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/levels/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit level completion with anti-cheat validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Levels"
                ],
                "summary": "Complete a level",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Level completion data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LevelCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Level completed successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.LevelCompletionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or cheating detected",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/talents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all talents and their levels for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Talents"
                ],
                "summary": "Get user talents",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User talents retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.UserTalent"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/talents/upgrade": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upgrade a talent using gold currency",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Talents"
                ],
                "summary": "Upgrade a talent",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer",
                        "description": "Bearer JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Talent upgrade request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TalentUpgradeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talent upgraded successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TalentUpgradeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request, insufficient gold, or max level",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/models.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.LevelCompletionRequest": {
            "type": "object",
            "required": [
                "finalHp",
                "levelId",
                "mapId",
                "timeSeconds"
            ],
            "properties": {
                "counterCount": {
                    "type": "integer",
                    "minimum": 0
                },
                "dashCount": {
                    "type": "integer",
                    "minimum": 0
                },
                "finalHp": {
                    "type": "number",
                    "minimum": 0
                },
                "levelId": {
                    "type": "string"
                },
                "mapId": {
                    "type": "string"
                },
                "timeSeconds": {
                    "type": "number"
                },
                "vulnerableKills": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "models.LevelCompletionResponse": {
            "type": "object",
            "properties": {
                "goldEarned": {
                    "type": "integer"
                },
                "isFirstCompletion": {
                    "type": "boolean"
                },
                "mapUnlocked": {
                    "type": "string"
                },
                "newBestTime": {
                    "type": "boolean"
                },
                "newTotalGold": {
                    "type": "integer"
                },
                "newTotalStars": {
                    "type": "integer"
                },
                "nextLevelUnlocked": {
                    "type": "string"
                },
                "starsEarned": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.TalentUpgradeRequest": {
            "type": "object",
            "required": [
                "talentId"
            ],
            "properties": {
                "talentId": {
                    "type": "string"
                }
            }
        },
        "models.TalentUpgradeResponse": {
            "type": "object",
            "properties": {
                "goldSpent": {
                    "type": "integer"
                },
                "newLevel": {
                    "type": "integer"
                },
                "newTotalGold": {
                    "type": "integer"
                },
                "nextLevelCost": {
                    "type": "integer"
                },
                "statBonus": {
                    "description": "e.g., \"+5% HP\", \"+3% Damage\"",
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "talentId": {
                    "type": "string"
                }
            }
        },
...
        "models.UserTalent": {
            "type": "object",
            "properties": {
                "currentLevel": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "talentId": {
                    "type": "string"
                },
                "upgradedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "PlayFabToken": {
            "description": "PlayFab session token for authentication",
            "type": "apiKey",
            "name": "X-PlayFab-SessionToken",
            "in": "header"
        }
    }
}